pyproject.toml
[project]
name = "lighton-ocr-cpu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.132.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "safetensors>=0.7.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.66.0",
    "transformers>=5.0.0",
    "uvicorn>=0.41.0",
]
